Added function to upload the database on memory
We store the tables as vector of objetcs (that corresponds to the raw of
the table).The vectors are stored in CryptoFile/Db/DatabaseValues and
can be accessed only by using the get method.

Each db::OriginalFile has stored inside the vector of his own db::Sections.
When a db::OriginalFile object is saved it automatically saves his
sections.